Stacked segments can be parsed, if made correctly
Note that previous message about stacked segments no longer applies--
was trying to parse a stacked segment that contained a poorly
made segment, so root cause was latter, not the fact it was stacked.
Stacked is handled just fine recursively.

Need to treat poorly made segments as errors and either stop, or
return what can be parsed so users can salvage what they can if
possible. Will otherwise go crazy trying to handle all manner of
edge cases that should really be errors (in the GUI)