Combination test and actual work
Push set of 676 URLs via SaveInteralURLFilters for colleague.
This is mainly to capture the workflow and document it.

Note that this function (the Save function) is still in draft
and is not in the package formally. Like other POST methods, it
needs quite a bit of infrastucture to make it "safe."

For now, just get the job done so others can move on with their
lives. Because 676 manual repetitions is 675 too many.